DESCRIPTION: This code is the DSP performance prediction code provided by professor. I refactored the code on 2021-11-04. I added code to run all activation functions easier
ToDo:
I need to test whether the code receives the same results. The unit test I made fails. This could be because the code slightly alters the results or there's some acceptable variance between attempts and my unit test is too sensitive. First I should prove where there's variance. I can create some code that creates a model, receive it's history, save it's history, then restart the code. If the results are identical after a repeat attempt, then there is no variance on first training run. If there's variance only after  I run the model, delete the model, then re-instantiated it during the same runtime with the same data, then I should save this data and compare it with the first run. If there is variance, then the test is too sensitive. I can create a test by examining the maximum variance and factoring this in my test case.  If this is too difficult, then I can use the outputted graph, with the previous graph to compare with previous figures of relu.

I need to fix a logical error for the runAll:  Not all model combinations run